+++ packages/react/src/themes/teams/components/Button/buttonVariables.ts
@@ -3,12 +3,14 @@ import { FontWeightProperty } from 'csstype'
 import { pxToRem } from '../../../../lib'
 
 export interface ButtonVariables {
+  padding: string
   height: string
   minWidth: string
   maxWidth: string
   borderRadius: string
-  paddingLeftRightValue: number
   contentFontWeight: FontWeightProperty
+  contentFontSize: string
+  contentLineHeight: string
 
   color: string
   colorHover: string
@@ -49,15 +51,24 @@ export interface ButtonVariables {
   textPrimaryColorHover: string
 
   boxShadow: string
+
+  sizeSmallContentFontSize: string
+  sizeSmallContentLineHeight: string
+  sizeSmallHeight: string
+  sizeSmallMinWidth: string
+  sizeSmallPadding: string
 }
 
 export default (siteVars: any): ButtonVariables => ({
+  padding: `0 ${pxToRem(20)}`,
   height: pxToRem(32),
   minWidth: pxToRem(96),
   maxWidth: pxToRem(280),
   borderRadius: siteVars.borderRadius,
+
+  contentFontSize: siteVars.fontSizes.medium,
   contentFontWeight: siteVars.fontWeightSemibold,
-  paddingLeftRightValue: 20,
+  contentLineHeight: siteVars.lineHeightMedium,
 
   color: siteVars.colors.grey[750],
   colorHover: siteVars.colors.grey[750],
@@ -98,4 +109,10 @@ export default (siteVars: any): ButtonVariables => ({
   textPrimaryColorHover: siteVars.colorScheme.brand.foregroundHover,
 
   boxShadow: siteVars.shadowLevel1,
+
+  sizeSmallContentFontSize: siteVars.fontSizes.small,
+  sizeSmallContentLineHeight: siteVars.lineHeightSmall,
+  sizeSmallHeight: pxToRem(24),
+  sizeSmallMinWidth: pxToRem(72),
+  sizeSmallPadding: `0 ${pxToRem(8)}`,
 })